+++ tests/functional/eyes_selenium/mobile/test_app_native.py
@@ -11,7 +11,7 @@ ANDROID_CAPS = {
     "automationName": "UiAutomator2",
     "platformName": "Android",
     "platformVersion": "6.0",
-    "appiumVersion": "1.13.0",
+    "appiumVersion": "1.17.1",
     "deviceName": "Android Emulator",
     "deviceOrientation": "portrait",
 }
@@ -22,9 +22,9 @@ IOS_CAPS = {
     "noReset": True,
     "browserName": "",
     "platformName": "iOS",
-    "platformVersion": "12.2",
+    "platformVersion": "13.4",
     "automationName": "XCUITest",
-    "appiumVersion": "1.13.0",
+    "appiumVersion": "1.17.1",
     "deviceName": "iPhone XR Simulator",
     "deviceOrientation": "portrait",
 }