+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [vNext]
+### Added
+- Concurrency version 2 protocol optimizations [Trello 2368](https://trello.com/c/0qi2c0jW)
 ### Updated
 - DOM snapshot script to *4.4.3* [Trello 2405](https://trello.com/c/wBy9nGQi)