+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [vNext]
+### Added
+- Concurrency version 2 protocol optimizations [Trello 2368](https://trello.com/c/0qi2c0jW)
 ### Fixed
 - Unable to capture multiple scrolled region screenshots [Trello 2397](https://trello.com/c/3RvCvojV)
 ### Updated